partically fixed negative total_found, docs and hits counter on huge indexes
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@3079 406a0c4d-033a-0410-8de8-e80135713968

+++ src/sphinx.cpp
@@ -17842,7 +17842,7 @@ void InfixBuilder_c<SIZE>::SaveEntries ( CSphWriter & wrDict )
 					sCur++;
 					sPrev++;
 				}
-				iKeepChars = (int)( sCur-(const BYTE*)sKey );
+				iKeepChars = (int)( sCur- ( const BYTE* ) sKey );
 
 				assert ( iKeepChars>=0 && iKeepChars<16 );
 				assert ( iChars-iKeepChars>=0 );
@@ -25248,7 +25248,7 @@ CSphQueryResultMeta::CSphQueryResultMeta ()
 }
 
 
-void CSphQueryResultMeta::AddStat ( const CSphString & sWord, int iDocs, int iHits, bool bExpanded )
+void CSphQueryResultMeta::AddStat ( const CSphString & sWord, int64_t iDocs, int64_t iHits, bool bExpanded )
 {
 	CSphString sFixed;
 	const CSphString * pFixed = &sWord;
